Ajout de l'interface graphique
J’ai crée l’interface mais elle n’est pas connecté encore complètement
(meme pas du tout)
je n’ai pas encore ajouter ton code sur le memento

+++ operateur.cpp
@@ -63,7 +63,7 @@ void FabriqueOperateur::supprimer(Operateur * Op)
         ++It;
         i++;
     }
-    if (i == OpTab.size()) throw OperateurException("Element inconnu");
+    if (i == (unsigned int)OpTab.size()) throw OperateurException("Element inconnu");
     OpTab.erase(OpTab.begin() + i);
     delete(Op);
 }
